third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/writing-mode-parsing-sideways-rl-001-expected.txt

This is a testharness.js-based test.
FAIL writing-mode: sideways-rl assert_equals: expected "sideways-rl" but got "horizontal-tb"
Harness: the test ran to completion.
